+++ docs_src/pagination/simple_pagination.py
@@ -16,19 +16,19 @@ class BlogEntry(edgy.Model):
 
 async def get_last_blogpost_page():
     # order by is required for paginators
-    paginator = Paginator(User.query.order_by("-created", "-id"), page_size=30)
+    paginator = Paginator(BlogEntry.query.order_by("-created", "-id"), page_size=30)
     return await paginator.get_page(-1), await paginator.get_amount_pages()
 
 
 async def get_blogpost_pages():
     # order by is required for paginators
-    paginator = Paginator(User.query.order_by("-created", "-id"), page_size=30)
+    paginator = Paginator(BlogEntry.query.order_by("-created", "-id"), page_size=30)
     return [page async for page in paginator.paginate()]
 
 
 async def search_blogpost(title: str, page: int):
     # order by is required for paginators
     paginator = Paginator(
-        User.query.filter(title__icontains=title).order_by("-created", "-id"), page_size=30
+        BlogEntry.query.filter(title__icontains=title).order_by("-created", "-id"), page_size=30
     )
     return await paginator.get_page(page), await paginator.get_amount_pages()