+++ edgy/contrib/pagination/__init__.py
@@ -1,4 +1,4 @@
-from .base import Paginator
-from .cursor import CursorPaginator
+from .base import BasePage, BasePaginator, Page, Paginator
+from .cursor import CursorPage, CursorPaginator
 
-__all__ = ["Paginator", "CursorPaginator"]
+__all__ = ["BasePage", "BasePaginator", "Page", "Paginator", "CursorPage", "CursorPaginator"]